Added visual buff icon for when "Start with Aghs" is enabled
Changed the granted modifier from regular aghs item, to the aghs upgrade that alch gives. The only difference is that the alch one comes with the visual buff. It still does not give any extra stats. I think its a nice touch for players to know from the start they have aghs (incase they were not paying attention to settings, or are newbs). The only thing that I would change if I knew how, would be to change the tooltip description when you hover over the buff. It says it grants stats, and it does'nt.