Survival balance
people just stay in fountain and waiting for the bonus. For the bonuses that can have effect at fountain(Cooldown spell amp), it has to be weak. For bonus that work in fight, should be stronger. Otherwise we should just make them troll combo with any global skills. Or, we make them stack much faster, but only grow when you are in enemies visions. for now, I'll just do some value balance.